Move the adapter into the scheduler
