+++ core/src/test/scala/com/mesosphere/usi/core/scalaapi/SchedulerAdapterTest.scala
@@ -33,7 +33,9 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       Goal.Running,
       RunSpec(
         resourceRequirements = List(ScalarRequirement.cpus(1), ScalarRequirement.memory(256)),
-        shellCommand = "sleep 3600")
+        shellCommand = "sleep 3600",
+        role = "marathon"
+      )
     )
     val specEvent = PodSpecUpdated(podSpec.id, None)
     val stateEvent = PodStatusUpdated(podSpec.id, None)
@@ -44,17 +46,17 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
   "SchedulerAdapter" when {
     "materialized as Source and Sink" should {
       "Propagate a snapshot to the underlying scheduler" in new SchedulerFixture {
-        SchedulerAdapter.asSourceAndSink(SpecsSnapshot(List(podSpec), Nil), scheduler)
+        Scheduler.asSourceAndSink(SpecsSnapshot(List(podSpec), Nil), scheduler)
         specInputQueue.pull().futureValue.value._1.podSpecs.head shouldEqual podSpec
       }
 
       "If snapshot is not provided push an empty snapshot" in new SchedulerFixture {
-        SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         specInputQueue.pull().futureValue.value._1 shouldEqual SpecsSnapshot.empty
       }
 
       "Complete the snapshot promise once it's available" in new SchedulerFixture {
-        val (snapshotF, _, _) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (snapshotF, _, _) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         specInputQueue.pull().futureValue
 
         val stateSnapshot = StateSnapshot(List(PodStatus(podSpec.id, Map.empty)), Nil, Nil, Nil)
@@ -65,7 +67,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Push the SpecUpdatedEvents downstream" in new SchedulerFixture {
-        val (_, _, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, _, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
 
         Source.repeat(specEvent).runWith(sink)
 
@@ -78,7 +80,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Push the events from scheduler to the provided Source" in new SchedulerFixture {
-        val (_, source, _) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, _) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
 
         stateOutputQueue.offer(StateSnapshot.empty -> Source.repeat(stateEvent))
 
@@ -86,7 +88,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Complete both sink and source when the event stream coming to scheduler is cancelled by the scheduler" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
 
         val stateStreamCompletionPromise = Promise[String]()
@@ -109,7 +111,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Complete both sink and source when the event stream coming to scheduler is cancelled by the client" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
 
         inside(specInputQueue.pull().futureValue) {
           case Some((_, innerSource)) => innerSource.runWith(Sink.ignore)
@@ -133,7 +135,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Complete both sink and source when the event stream outgoing from scheduler is cancelled by the scheduler" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.runWith(Sink.ignore) // consume all incoming spec updates
 
@@ -158,7 +160,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Complete both sink and source when the event stream outgoing from scheduler is cancelled by the client" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.runWith(Sink.ignore) // consume all incoming spec updates
 
@@ -183,7 +185,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Fail both sink and source when the failure occurred in the upstream" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.runWith(Sink.ignore) // consume all incoming spec updates
 
@@ -211,7 +213,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Stop the sink when the failure occurred in the downstream" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.runWith(Sink.ignore) // consume all incoming spec updates
 
@@ -240,7 +242,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Stop both sink and source when the failure occurred in the scheduler input" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.map(_ => throw new Exception("Boom!")).runWith(Sink.ignore) // explosion in the scheduler
 
@@ -277,7 +279,7 @@ class SchedulerAdapterTest extends AkkaUnitTest with Inside {
       }
 
       "Stop sink and fail source when the failure occurred in the scheduler output" in new SchedulerFixture {
-        val (_, source, sink) = SchedulerAdapter.asSourceAndSink(SpecsSnapshot.empty, scheduler)
+        val (_, source, sink) = Scheduler.asSourceAndSink(SpecsSnapshot.empty, scheduler)
         val (_, innerSource) = specInputQueue.pull().futureValue.value
         innerSource.runWith(Sink.ignore)
